requirements.txt
pandas
plotly  
numpy==1.26.4
dvc
matplotlib
notebook
scikit-learn==1.3.0 # phiên bản này phù hợp với xgboost
python-box==6.0.2
pyYAML
types-PyYAML
seaborn
xgboost
lightgbm 
imbalanced-learn
scipy
zzzz_mylib_23_4==0.53
-e .